Fix nightly build + move EmbreeIntersector impl to .cpp
